Support the `FldName` namespace introduced in GHC 9.8
GHC 9.8 now puts record names into the new `FldName` namespace instead of
`VarName`. This patch updates `th-desugar` accordingly.

One consequence of supporting `FldName` properly is that the
`lookupValueNameWithLocals` function will no longer reify field selectors for
record names when `NoFieldSelectors` is set. This is a desirable property, as
this mirrors the behavior of the `lookupValueName` function in Template
Haskell, thereby fixing #183. I've made a note of this change in behavior in
the `CHANGES.md` file.

Note that the way we check for `NoFieldSelectors` is somewhat unsatisfactory.
As Richard notes in
https://github.com/goldfirere/th-desugar/issues/183#issuecomment-1517792696,
`th-desugar` can only check for language extensions at the use sites of
declarations, whereas it would be better to check for language extensions at
definition sites.  Unfortunately, Template Haskell does not offer a way to do
the latter, so we have no choice but to do the former.

This isn't really a new thing, since we also make the same compromise when
checking for `MonadFailDesugaring` when desugaring partial pattern matches in
`do` notation. As a result, I've opted to simpy make a note of this limitation
in the `th-desugar` `README`.

Checks off one box in #182.